Backend/Backend/src/test/java/ca/mcgill/ecse321/backend/service/InternshipServiceTest.java
package ca.mcgill.ecse321.backend.service;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertNotNull;
import static org.junit.Assert.assertNull;
import static org.mockito.ArgumentMatchers.anyString;
import static org.mockito.Mockito.when;

import ca.mcgill.ecse321.backend.dao.InternshipRepository;
import ca.mcgill.ecse321.backend.dto.InternshipDto;
import ca.mcgill.ecse321.backend.model.*;
import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.invocation.InvocationOnMock;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.security.core.context.SecurityContextHolder;
import org.springframework.security.core.userdetails.User;
import org.springframework.security.crypto.password.PasswordEncoder;
import org.springframework.security.test.context.support.WithAnonymousUser;
import org.springframework.security.test.context.support.WithMockUser;
import org.springframework.stereotype.Service;
import org.springframework.test.context.junit4.SpringRunner;

import ca.mcgill.ecse321.backend.dao.StudentRepository;

@RunWith(SpringRunner.class)
@SpringBootTest
public class InternshipServiceTest {

    @Autowired
    private InternshipRepository internshipRepository;

    @Autowired
    private InternshipService service;

    private ApplicationForm applicationForm;
    private

    private Course mockCourse;

    @Test
    public void createInternship() {
    }

    @Test
    public void findByIdAndStudentStudentID() {
    }

    @Test
    public void findByIdAndStudent() {
    }

    @Test
    public void toDto() {
    }

}
